+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Release 0.1.2 (Sun Jul 2022)
+
+  * Breaking Changes
+    * In save/savez/savecsv, the order of arguments is changed.
+
+  * Major Features and Improvements
+    * Improved savecsv/2, ablity to handle %Nx.tensor
+    * Added to_csv/1, convert %Npy{}/Nx.tensor to csv string.
+    * Added load!/2.
+
 ## Release 0.1.1 (Jul 11 2022)
 
   * Major Features and Improvements